Comment update Sample app upgrade and updates
