+++ samples/Sample.Core/Sample.Core.csproj
@@ -34,8 +34,8 @@
       <HintPath>..\packages\RazorEngine.3.8.2\lib\net45\RazorEngine.dll</HintPath>
       <Private>True</Private>
     </Reference>
-    <Reference Include="RazorMailer.Core, Version=0.9.9.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\RazorMailer.0.9.9\lib\net452\RazorMailer.Core.dll</HintPath>
+    <Reference Include="RazorMailer.Core, Version=0.9.10.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\RazorMailer.0.9.10\lib\net452\RazorMailer.Core.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="System" />
@@ -54,7 +54,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Email\Models\WelcomeModel.cs" />
-    <Compile Include="Email\SampleMailer.cs" />
+    <Compile Include="Email\Mailer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
