+++ samples/Sample.Core/packages.config
@@ -2,5 +2,5 @@
 <packages>
   <package id="Microsoft.AspNet.Razor" version="3.0.0" targetFramework="net452" />
   <package id="RazorEngine" version="3.8.2" targetFramework="net452" />
-  <package id="RazorMailer" version="0.9.9" targetFramework="net452" />
+  <package id="RazorMailer" version="0.9.12" targetFramework="net452" />
 </packages>
