+++ samples/Sample.Web/packages.config
@@ -24,6 +24,8 @@
   <package id="Modernizr" version="2.6.2" targetFramework="net452" />
   <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net452" />
   <package id="Owin" version="1.0" targetFramework="net452" />
+  <package id="RazorEngine" version="3.8.2" targetFramework="net452" />
+  <package id="RazorMailer" version="0.9.12" targetFramework="net452" />
   <package id="Respond" version="1.2.0" targetFramework="net452" />
   <package id="SimpleInjector" version="3.1.2" targetFramework="net452" />
   <package id="SimpleInjector.Integration.Web" version="3.1.2" targetFramework="net452" />
